+++ src/test/java/persistence/sql/ddl/DDLGeneratorTest.java
@@ -5,15 +5,15 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class MySqlDDLGeneratorTest {
+class DDLGeneratorTest {
 
     @Test
     void createPersonCreateQuery() {
         // given
-        MySqlDDLGenerator mySqlDdlGenerator = new MySqlDDLGenerator();
+        DDLGenerator ddlGenerator = new DDLGenerator();
 
         // when
-        String createSql = mySqlDdlGenerator.generateCreate(Person.class);
+        String createSql = ddlGenerator.generateCreate(Person.class);
 
         // then
         assertThat(createSql).isEqualTo("CREATE TABLE users (" +
@@ -26,10 +26,10 @@ class MySqlDDLGeneratorTest {
     @Test
     void createPersonDropQuery() {
         // given
-        MySqlDDLGenerator mySqlDdlGenerator = new MySqlDDLGenerator();
+        DDLGenerator ddlGenerator = new DDLGenerator();
 
         // when
-        String dropSql = mySqlDdlGenerator.generateDrop(Person.class);
+        String dropSql = ddlGenerator.generateDrop(Person.class);
 
         // then
         assertThat(dropSql).isEqualTo("DROP TABLE users;");
