+++ src/test/java/persistence/sql/ddl/table/TableTest.java
@@ -2,21 +2,21 @@ package persistence.sql.ddl.table;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.ddl.MySqlDDLGenerator;
+import persistence.sql.ddl.DDLGenerator;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.ThrowableAssert.catchThrowable;
 
-class MySqlTableTest {
+class TableTest {
 
     @Test
     @DisplayName("테이블에 @Id 가 없을 경우 예외가 발생한다")
     void failByNotFoundIdAnnotation() {
         // given
-        MySqlDDLGenerator mySqlDDLGenerator = new MySqlDDLGenerator();
+        DDLGenerator DDLGenerator = new DDLGenerator();
 
         // when
-        Throwable throwable = catchThrowable(() -> mySqlDDLGenerator.generateCreate(PersonNotHaveIdAnnotation.class));
+        Throwable throwable = catchThrowable(() -> DDLGenerator.generateCreate(PersonNotHaveIdAnnotation.class));
 
         // then
         assertThat(throwable)
